5309: Confirmation dialog for GeoStory pending changes
