+++ web/client/localConfig.json
@@ -277,7 +277,7 @@
             },
             "FeedbackMask"
         ],
-        "desktop": [ "Details",
+        "desktop": [ "Details", "Snapshot",
           {
             "name": "Map",
             "cfg": {
